server/requirements.txt
asgiref==3.9.1
boto3==1.39.10
botocore==1.39.10
dj-database-url==3.0.1
django==5.2.4
django-cors-headers==4.7.0
django-otp==1.6.1
django-storages==1.14.6
django-tinymce==4.1.0
djangorestframework==3.16.0
jmespath==1.0.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==8.2
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
